+++ node/tests/api/notebook/openNotebook.test.ts
@@ -37,28 +37,36 @@ describe(pathname, async () => {
     const validate_payload = schema_payload.constructValidateFuction();
     const validate_response = schema_response.constructValidateFuction();
 
+    const notebook_name = "openNotebook"
     testKernelAPI<openNotebook.IPayload, openNotebook.IResponse>({
         name: "main",
         payload: {
             data: {
-                notebook: "20211226090932-5lcq56f", // 思源筆記用戶指南
+                notebook: "", // 将使用新创建的笔记本的 ID
             },
             validate: validate_payload,
-            test: async () => {
-                await client.client.openNotebook({
-                    notebook: "20211226090932-5lcq56f",
+            test: async (payload) => {
+                /* 新建一个笔记本以进行测试 */
+                const response =  await client.client.createNotebook({
+                    name: notebook_name,
                 });
-            }
+                payload.notebook = response.data.notebook.id;
+            },
         },
         request: (payload) => client.client.openNotebook(payload!),
         response: {
             validate: validate_response,
-            test: async () => {
+            test: async (_response, payload) => {
                 test("test the status of notebook", async () => {
                     const response = await client.client.getNotebookConf({
-                        notebook: "20211226090932-5lcq56f",
+                        notebook: payload!.notebook,
                     });
                     expect(response.data.conf.closed).toBeFalsy;
+
+                    /* 删除新建的笔记本 */
+                    await client.client.removeNotebook({
+                        notebook: payload!.notebook,
+                    });
                 });
             },
         },